pulled ReportBatch, ReportBatchIterator, ReportFileList classes up to module top level
